+++ frontend/workflows/serverexperimentation/src/start-experiment.tsx
@@ -12,46 +12,9 @@ import {
 import { useDataLayout } from "@clutch-sh/data-layout";
 import type { WizardChild } from "@clutch-sh/wizard";
 import { Wizard, WizardStep } from "@clutch-sh/wizard";
-import { FormControl, FormControlLabel, FormLabel, Radio, RadioGroup } from "@material-ui/core";
 import * as yup from "yup";
 
-interface RadioControlItem {
-  label: string;
-  value: string;
-}
-
-interface RadioControlProps {
-  name: string;
-  label: string;
-  items: RadioControlItem[];
-  onChange: (value: string) => void;
-}
-
-const RadioControl: React.FC<RadioControlProps> = ({ name, label, items, onChange }) => {
-  return (
-    <FormControl key={name}>
-      <FormLabel component="legend">Upstream Cluster Type</FormLabel>
-      <RadioGroup
-        aria-label={label}
-        name={name}
-        defaultValue={items[0].value}
-        onChange={e => onChange(e.target.value)}
-      >
-        {items &&
-          items.map(item => {
-            return (
-              <FormControlLabel
-                key={item.value}
-                value={item.value}
-                control={<Radio />}
-                label={item.label}
-              />
-            );
-          })}
-      </RadioGroup>
-    </FormControl>
-  );
-};
+import { RadioControl } from "./radio-control";
 
 const faultInjectionTypeItems = [
   {
